requirements.txt
zxcvbn-python>=4.4.24
streamlit>=1.36
pandas>=2.0
pytest>=8.0
